Solution Day 13 pt. 2
This is a naive cheat solution. It probably would have been easier to
solve it straightforwardly, but I thought this might be quicker.

Essentially, the program plays the game over and over, noting where the
paddle should have been to prevent last round's loss and making sure it
moves to that position in the next round. It repeats this until all
blocks are broken.

Solving for the solution initially is very slow since the blocks get
quite sparse toward the end, and the round gets longer as you break more
blocks before losing.

Could vastly improve this strategy by taking a snap shot of the computer
and the game state at the last known correct position of the paddle, then
rewinding to that snapshot after the next paddle target is acquired.